intrinsics attribute opt-out list proposal.
Summary:
This patch introduces an opt-out attribute list for intrinsics, discussed in D65377.
This should make it easier to annotate intrinsics with attributes that apply to most of them.
Right now I added NoFree and NoSync. If this looks ok, I'll also replace "Throws" IntrinsicProperty with IntrNoUnwind
and make it default.

Note: This is just an attempt, I also intend to send a RFC on the list.

Reviewers: jdoerfert, arsenm

Subscribers: wdng, llvm-commits, uenoku

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D70365