[IR] Intrinsics attribute opt-out list proposal.
Summary: This patch introduces an opt-out attribute list for intrinsics, discussed in D65377.
This should make it easier to annotate intrinsics with attributes that apply to most of them.
Right now I added NoFree and NoSync. If this looks ok, I'll also replace "Throws" IntrinsicProperty with IntrNoUnwind
and make it default.