[OpenMPOpt] ICV tracking for calls
This introduces the ability to get the unique ICV value for a call, if
possible.
There is a slight change in the aproach. We are now tracking(mapping)
Values to instructions in a DenseMap instead of using SmallSetVector<ICVValue>.
In addition to that we are now going through all calls looking for the
first one that has tracked value, instead of trying to reach certain
setter. The first approach could possibly even map the wrong value(setter),
due to order in which setters were collected.

Differential Revision: https://reviews.llvm.org/D85544